backend/requirements/common.txt
fastapi==0.115.12
sqlalchemy==2.0.40
pydantic==2.11.4
pydantic_settings==2.9.1
python-jose==3.4.0
requests==2.32.3
python-dotenv==1.0.0
uvicorn==0.34.2
psycopg2-binary==2.9.10
pydantic[email]==2.11.4
uvicorn[standard]==0.34.2